+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## HEAD (Unreleased)
 
 - Disable last-applied-configuration annotation for replaced CRDs (https://github.com/pulumi/pulumi-kubernetes/pull/1868)
+- Fix Provider config diffs (https://github.com/pulumi/pulumi-kubernetes/pull/1869)
 - Fix replace for named resource using server-side diff (https://github.com/pulumi/pulumi-kubernetes/pull/1870)
 
 ## 3.14.0 (January 12, 2022)